# app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stock Investment Analyzer API
  description: |
    株式投資分析システムのAPI仕様書

    このAPIは以下の機能を提供します：
    - 株式データの取得と管理
    - バルクデータ処理
    - システム監視とヘルスチェック

  version: 1.0.0
  contact:
    name: Stock Investment Analyzer Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: 開発環境
  - url: http://localhost:8000/api/v1
    description: 開発環境 (v1)
...
paths:
  /api/stocks:
    get:
      tags:
        - 株価データ
      summary: 株式データ一覧取得
      description: 登録されている株式データの一覧を取得します
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Stock'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/stocks/{stock_id}:
    get:
      tags:
        - 株価データ
      summary: 株式データ詳細取得
      description: 指定された株式IDの詳細データを取得します
      parameters:
        - name: stock_id
          in: path
          required: true
          description: 株式ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Stock'
        '404':
          description: 株式データが見つかりません
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/stock-master/stocks:
    get:
      tags:
        - 銘柄マスター
      summary: 株式マスターデータ取得
      description: 株式マスターデータの一覧を取得します
      parameters:
        - name: limit
          in: query
          description: 取得件数の上限
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: 取得開始位置
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StockMaster'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/system/database/connection:
    get:
      tags:
        - システム監視
      summary: データベース接続確認
      description: データベースへの接続状況を確認します
      responses:
        '200':
          description: 接続正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: connected
                  database:
                    type: string
                    example: "postgresql"
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/system/external-api/connection:
    get:
      tags:
        - システム監視
      summary: 外部API接続確認
      description: 外部APIへの接続状況を確認します
      responses:
        '200':
          description: 接続正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: connected
                  apis:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        status:
                          type: string
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  responses:
...
    InternalServerError:
      description: 内部サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: サービス利用不可
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Stock:
      type: object
      properties:
        id:
          type: integer
          description: 株式ID
          example: 1
        symbol:
          type: string
          description: 株式コード
          example: "7203"
        name:
          type: string
          description: 株式名
          example: "トヨタ自動車"
        market:
          type: string
          description: 市場
          example: "東証プライム"
        created_at:
          type: string
          format: date-time
          description: 作成日時
        updated_at:
          type: string
          format: date-time
          description: 更新日時
...
    StockMaster:
      type: object
      properties:
        code:
          type: string
          description: 株式コード
          example: "7203"
        name:
          type: string
          description: 株式名
          example: "トヨタ自動車"
        market:
          type: string
          description: 市場
          example: "東証プライム"
        sector:
          type: string
          description: 業種
          example: "輸送用機器"
        industry:
          type: string
          description: 業界
          example: "自動車"
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          description: エラーメッセージ
        error:
          type: object
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: 詳細エラーメッセージ
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 総件数
          example: 1000
        limit:
          type: integer
          description: 取得件数上限
          example: 100
        offset:
          type: integer
          description: 取得開始位置
          example: 0
        has_next:
          type: boolean
          description: 次のページが存在するか
          example: true
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API認証キー

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン
security:
  - ApiKeyAuth: []
  - BearerAuth: []